ci: fix bug and add sale product

+++ server/src/app/controllers/OrderController.js
@@ -2,6 +2,7 @@ const ip = require("ip");
 const CartItem = require("../models/CartItem");
 const ItemOrder = require("../models/ItemOrder");
 const Order = require("../models/Order");
+const Product = require("../models/Product");
 
 class OrderController {
   getAllOrders(req, res) {
@@ -24,73 +25,92 @@ class OrderController {
       });
   }
 
-  create(req, res, next) {
-    var ipAddr = ip.address();
+  async createPaymentUrl(req, res, next) {
+    const mapLoop = async () => {
+      // check xem có itemCart nào không đủ số lượng
+      const promises = await Object.values(req.body.cartItems).map((item) => {
+        const result = new Promise((resolve, reject) => {
+          Product.findById(item[0].idProduct._id).then((product) => {
+            if (product.stock < item[0].quantity)
+              return resolve(
+                "Sản phẩm " + product.title + " không đủ số lượng"
+              );
+            return resolve(true);
+          });
+        });
+        return result;
+      });
+      const results = await Promise.all(promises);
 
-    var config = require("../../config/default.json");
-    var dateFormat = require("date-format");
+      // nếu có sản phẩm nào không đủ số lượng thì trả về thông báo, còn đủ thì trả về urlVNPay
+      if (results.every((item) => item === true)) {
+        var ipAddr = ip.address();
+        var config = require("../../config/default.json");
+        var dateFormat = require("date-format");
+        var tmnCode = config.vnp_TmnCode;
+        var secretKey = config.vnp_HashSecret;
+        var vnpUrl = config.vnp_Url;
+        var returnUrl = `${process.env.VNP_RETURNURL}order/vnpay_return`;
+        var date = new Date();
+        var createDate =
+          date.getFullYear() +
+          "" +
+          ("0" + (date.getMonth() + 1)).slice(-2) +
+          "" +
+          ("0" + date.getDate()).slice(-2) +
+          "" +
+          ("0" + date.getHours()).slice(-2) +
+          "" +
+          ("0" + date.getMinutes()).slice(-2) +
+          "" +
+          ("0" + date.getSeconds()).slice(-2);
+        var orderId = dateFormat(date, "HHmmss");
+        var amount = req.body.amount;
+        // var bankCode = "NCB";
+        var bankCode = "";
+        var orderInfo = "Thanh toan don hang";
+        var orderType = "other";
+        var locale = "vn";
+        if (locale === null || locale === "") {
+          locale = "vn";
+        }
+        var currCode = "VND";
+        var vnp_Params = {};
+        vnp_Params["vnp_Version"] = "2.1.0";
+        vnp_Params["vnp_Command"] = "pay";
+        vnp_Params["vnp_TmnCode"] = tmnCode;
+        vnp_Params["vnp_Locale"] = locale;
+        vnp_Params["vnp_CurrCode"] = currCode;
+        vnp_Params["vnp_TxnRef"] = orderId;
+        vnp_Params["vnp_OrderInfo"] = orderInfo;
+        vnp_Params["vnp_OrderType"] = orderType;
+        vnp_Params["vnp_Amount"] = amount * 100;
+        vnp_Params["vnp_ReturnUrl"] = returnUrl;
+        vnp_Params["vnp_IpAddr"] = ipAddr;
+        vnp_Params["vnp_CreateDate"] = createDate;
+        if (bankCode !== null && bankCode !== "") {
+          vnp_Params["vnp_BankCode"] = bankCode;
+        }
+        vnp_Params = sortObject(vnp_Params);
+        var querystring = require("qs");
+        var signData = querystring.stringify(vnp_Params, { encode: false });
+        var crypto = require("crypto");
+        var hmac = crypto.createHmac("sha512", secretKey);
+        // var signed = hmac.update(new Buffer(signData, "utf-8")).digest("hex");
+        var signed = hmac
+          .update(new Buffer.from(signData, "utf-8"))
+          .digest("hex");
+        vnp_Params["vnp_SecureHash"] = signed;
+        vnpUrl += "?" + querystring.stringify(vnp_Params, { encode: false });
+        if (!req.body.amount) return res.json({ message: "Missing amount" });
+        res.json({ data: vnpUrl });
+      } else {
+        const itemDifferent = results.filter((item) => item !== true);
+        res.status(400).json({ message: itemDifferent[0] });
+      }
+    };
 
-    var tmnCode = config.vnp_TmnCode;
-    var secretKey = config.vnp_HashSecret;
-    var vnpUrl = config.vnp_Url;
-    var returnUrl = `${process.env.VNP_RETURNURL}order/vnpay_return`;
-
-    var date = new Date();
-
-    var createDate =
-      date.getFullYear() +
-      "" +
-      ("0" + (date.getMonth() + 1)).slice(-2) +
-      "" +
-      ("0" + date.getDate()).slice(-2) +
-      "" +
-      ("0" + date.getHours()).slice(-2) +
-      "" +
-      ("0" + date.getMinutes()).slice(-2) +
-      "" +
-      ("0" + date.getSeconds()).slice(-2);
-    var orderId = dateFormat(date, "HHmmss");
-    var amount = req.body.amount;
-    // var bankCode = "NCB";
-    var bankCode = "";
-
-    var orderInfo = "Thanh toan don hang";
-    var orderType = "other";
-    var locale = "vn";
-    if (locale === null || locale === "") {
-      locale = "vn";
-    }
-    var currCode = "VND";
-    var vnp_Params = {};
-    vnp_Params["vnp_Version"] = "2.1.0";
-    vnp_Params["vnp_Command"] = "pay";
-    vnp_Params["vnp_TmnCode"] = tmnCode;
-    vnp_Params["vnp_Locale"] = locale;
-    vnp_Params["vnp_CurrCode"] = currCode;
-    vnp_Params["vnp_TxnRef"] = orderId;
-    vnp_Params["vnp_OrderInfo"] = orderInfo;
-    vnp_Params["vnp_OrderType"] = orderType;
-    vnp_Params["vnp_Amount"] = amount * 100;
-    vnp_Params["vnp_ReturnUrl"] = returnUrl;
-    vnp_Params["vnp_IpAddr"] = ipAddr;
-    vnp_Params["vnp_CreateDate"] = createDate;
-    if (bankCode !== null && bankCode !== "") {
-      vnp_Params["vnp_BankCode"] = bankCode;
-    }
-
-    vnp_Params = sortObject(vnp_Params);
-
-    var querystring = require("qs");
-    var signData = querystring.stringify(vnp_Params, { encode: false });
-    var crypto = require("crypto");
-    var hmac = crypto.createHmac("sha512", secretKey);
-    // var signed = hmac.update(new Buffer(signData, "utf-8")).digest("hex");
-    var signed = hmac.update(new Buffer.from(signData, "utf-8")).digest("hex");
-    vnp_Params["vnp_SecureHash"] = signed;
-    vnpUrl += "?" + querystring.stringify(vnp_Params, { encode: false });
-
-    if (!req.body.amount) return res.json({ message: "Missing amount" });
-    res.json({ data: vnpUrl });
+    mapLoop();
   }
 
   vnpayReturn(req, res, next) {
@@ -148,8 +168,9 @@ class OrderController {
         function (err, cartItems) {
           if (err || cartItems.length === 0)
             return res.status(400).json({ error: err });
-          const grouped = {};
+
           // gom nhóm các sản phẩm giống nhau
+          const grouped = {};
           cartItems.forEach(function (a) {
             const itemOrder = new ItemOrder({
               product: a.idProduct,
@@ -165,24 +186,49 @@ class OrderController {
               grouped[a.idProduct._id + a.size + a.color] = [itemOrder];
             }
           });
+
           // tạo order
           const order = new Order({
             idAuth,
             order: grouped,
           });
-          //lưu order
-          order
-            .save()
-            .then(() => {
-              // res.status(200).json({ order });
-              // lưu order thành công thì xóa cartItem
-              CartItem.deleteMany({
-                idAuth,
-              }).then(() => res.status(200).json({ message: "success" }));
-            })
-            .catch((err) => {
-              res.status(400).json({ error: err });
+
+          try {
+            //lưu order
+            order.save().then(async (orderResult) => {
+              // update số lượng sản phẩm
+              const promises = await Object.values(orderResult.order).map(
+                (item) => {
+                  const result = new Promise((resolve) => {
+                    Product.findOneAndUpdate(
+                      { _id: item[0].product._id },
+                      { $inc: { stock: -item[0].quantity } },
+                      { new: true }
+                    )
+                      .then(resolve(true))
+                      .catch(resolve(false));
+                  });
+
+                  return result;
+                }
+              );
+
+              const results = await Promise.all(promises);
+
+              if (results.every((item) => item === true)) {
+                // lưu order thành công thì xóa cartItem
+                CartItem.deleteMany({
+                  idAuth,
+                }).then(() => {
+                  res.json({ message: "Order success" });
+                });
+              } else {
+                res.status(400).json({ message: "Order fail" });
+              }
             });
+          } catch (err) {
+            return res.status(400).json({ error: err });
+          }
         }
       );
     });
